Removed launchers for the exception testing and merged both exception scenarios into ExceptionTest
