--- src/test/java/at/fhhagenberg/app/exception_handling/ExceptionLauncher.java
@@ -1,9 +1,0 @@
-package at.fhhagenberg.app.exception_handling;
-
-import javafx.application.Application;
-
-public class ExceptionLauncher {
-    public static void main(String[] args) {
-        Application.launch(ECCAppExceptionTest.class, args);
-    }
-}
