--- src/test/java/at/fhhagenberg/app/exception_handling/NullServiceLauncher.java
@@ -1,9 +1,0 @@
-package at.fhhagenberg.app.exception_handling;
-
-import javafx.application.Application;
-
-public class NullServiceLauncher {
-    public static void main(String[] args) {
-        Application.launch(ECCAppNullServiceTest.class, args);
-    }
-}
